+++ models/PerclCommand.ts
@@ -33,6 +33,7 @@ import { SetTalk } from './SetTalk';
 import { Sms } from './Sms';
 import { StartRecordCall } from './StartRecordCall';
 import { TerminateConference } from './TerminateConference';
+import { TranscribeUtterance } from './TranscribeUtterance';
 import { Unpark } from './Unpark';
 import { HttpFile } from '../http/http';