+++ models/SetTalk.ts
@@ -34,6 +34,7 @@ import { SetTalkAllOf } from './SetTalkAllOf';
 import { Sms } from './Sms';
 import { StartRecordCall } from './StartRecordCall';
 import { TerminateConference } from './TerminateConference';
+import { TranscribeUtterance } from './TranscribeUtterance';
 import { Unpark } from './Unpark';
 import { HttpFile } from '../http/http';