+++ models/TerminateConference.ts
@@ -34,6 +34,7 @@ import { SetTalk } from './SetTalk';
 import { Sms } from './Sms';
 import { StartRecordCall } from './StartRecordCall';
 import { TerminateConferenceAllOf } from './TerminateConferenceAllOf';
+import { TranscribeUtterance } from './TranscribeUtterance';
 import { Unpark } from './Unpark';
 import { HttpFile } from '../http/http';